+++ gdaps/frontend/__init__.py
@@ -1,9 +1,10 @@
 import logging
 
 from gdaps.exceptions import PluginError
+from gdaps.frontend.api import IPackageManager
 from .api import IFrontendEngine, IPackageManager
 from .conf import frontend_settings
-from .engines import vue
+import gdaps.frontend.engines
 
 default_app_config = "gdaps.frontend.apps.FrontendConfig"
 logger = logging.getLogger(__file__)
@@ -12,10 +13,11 @@ __current_engine: IFrontendEngine or None = None
 
 
 def current_engine() -> IFrontendEngine:
-    """Returns the current frontend engine."""
+    """Returns the current frontend engine.
 
-    # while this function actually returns a plugin class that mocks IFrontendEngine,
-    # it's ok to say it returns one, ducktype style.
+    It caches the result for faster access.
+    """
+
     global __current_engine
     if __current_engine:
         return __current_engine
@@ -24,7 +26,7 @@ def current_engine() -> IFrontendEngine:
         if engine.name == frontend_settings.FRONTEND_ENGINE:
             __current_engine = engine
             return engine
-
-    raise PluginError(
-        "No frontend engine is selected. Please select one in settings.py using GDAPS['FRONTEND_ENGINE']"
-    )
+    else:
+        raise PluginError(
+            "No frontend engine is selected. Please select one in settings.py using GDAPS['FRONTEND_ENGINE']"
+        )