+++ gdaps/frontend/management/commands/syncplugins.py
@@ -1,21 +1,18 @@
 import logging
 
 from gdaps.frontend import current_engine
-from gdaps.pluginmanager import PluginManager
-from gdaps.management.commands.syncplugins import Command as SyncPluginsCommand
+from gdaps.management.commands.syncplugins import Command as BaseSyncPluginsCommand
 
 logger = logging.getLogger(__name__)
 
 
-class Command(SyncPluginsCommand):
+class Command(BaseSyncPluginsCommand):
     """This command adds frontend capabilities to GDAPS' base syncplugins command.."""
 
-    help = "Synchronizes all plugins into the database, and includes frontends."
+    help = "Synchronizes all plugins into the database, and installs their frontend packages."
 
     def handle(self, *args, **options) -> None:
         """Synchronizes all found plugins into the database."""
 
         super().handle(*args, **options)
-        current_engine().update_plugins_list(
-            [app.path for app in PluginManager.plugins()]
-        )
+        current_engine().update_plugins_list()