+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- allow using variables in template file/dir names
 
 ## [0.4.0] - 2019-10-12
 - Change API to an easier one, following Marty Alchin's Simple Plugin Framework